requirements.txt
config-path
markdown
markdown_strings
textual
textual[dev]

jinja2
pytest
pytest-rich/
syrupy
